Allow for controlled field updates
This update fixes and improves the updateField mechanism, allowing for the
ability to override fields.

A new method, `updateFields`, was created to enable the ability to update
multiple fields at once.

Lastly, a new method `unsafeOverrideFields` was created to enable the ability
forcefully change the fields to entirely new values.